Add a new level of folders to Forms folders to organize the files
