+++ src/app/routes/index.jsx
@@ -3,9 +3,9 @@ import { Routes, Route } from 'react-router-dom'
 import Home from '../pages/home/homeContainer'
 import About from '../pages/dummyAbout/about'
 import BackOfficeContainer from '../../components/BackOffice/BackOfficeContainer'
-import HomeFormContainer from '../../components/Forms/HomeFormContainer'
-import LoginFormContainer from '../../components/Forms/LoginFormContainer'
-import RegisterFormContainer from '../../components/Forms/RegisterFormContainer'
+import HomeFormContainer from '../../components/Forms/HomeForm/HomeFormContainer'
+import LoginFormContainer from '../../components/Forms/LoginForm/LoginFormContainer'
+import RegisterFormContainer from '../../components/Forms/RegisterForm/RegisterFormContainer'
 import NewsContainer from '../../components/News/NewsContainer'
 import NewsByIdContainer from '../../components/News/NewByIdContainer'
 import MyProfileContainer from '../../components/MyProfile/MyProfileContainer'
