+++ src/components/Forms/LoginForm/LoginForm.jsx
@@ -5,7 +5,7 @@ import {
   Container, CssBaseline, Box, Typography, Grid, Button,
 } from '@mui/material'
 import { Link } from 'react-router-dom'
-import FormInputField from '../Layout/FormInputField'
+import FormInputField from '../../Layout/FormInputField'
 import './loginForm.css'
 
 const LoginForm = ({
